requirements.txt
asgiref==3.2.10
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.9
distlib==0.3.4
Django==3.1.1
django-cors-headers==3.6.0
django-seed==0.2.2
django-storages==1.10.1
django-upgrade==1.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
dropbox==10.6.0
Faker==6.0.0
filelock==3.4.0
gunicorn==20.0.4
idna==2.10
Pillow==8.0.1
platformdirs==2.4.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
stripe==2.56.0
text-unidecode==1.3
tokenize-rt==4.2.1
urllib3==1.26.2
virtualenv==20.10.0
